requirements.txt
streamlit==1.38.0
sqlalchemy==2.0.35
psycopg2==2.9.9
pandas==2.2.3
python-dotenv
chromadb
openai==1.47.1
langchain-community==0.3.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.5
